+++ gdscript/CHANGELOG.md
@@ -6,6 +6,7 @@ Each release is delayed usually for 2 working days as it awaits JetBrain's appro
 
 + Comment indentation [#198](https://gitlab.com/IceExplosive/gdscript/-/issues/198)
 + Static method reference [#183](https://gitlab.com/IceExplosive/gdscript/-/issues/183)
++ Formatter - apply only "maximum lines" as described [#181](https://gitlab.com/IceExplosive/gdscript/-/issues/181)
 
 ### 2.1.8 (8.2.2024)